chore: remove unused components
