+++ src/components/frame/Extensions/layouts/index.ts
@@ -1,4 +1,3 @@
 export { default as ColonyLayout } from './ColonyLayout.tsx';
 export { default as MainLayout } from './MainLayout.tsx';
 export { default as UserNavigationWrapper } from './partials/UserNavigationWrapper/index.ts';
-export { default as MainSidebar } from './MainSidebar.tsx';
