--- src/components/frame/Extensions/layouts/partials/DashboardContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DashboardContent.tsx';
