+++ src/components/frame/Extensions/layouts/types.ts
@@ -1,6 +1,5 @@
 import { type ReactNode } from 'react';
 
-import { type NavigationSidebarProps } from '~v5/frame/NavigationSidebar/types.ts';
 import type { CalamityBannerItemProps } from '~v5/shared/CalamityBanner/types.ts';
 
 export interface UseCalamityBannerInfoReturnType {
@@ -9,9 +8,6 @@ export interface UseCalamityBannerInfoReturnType {
 }
 
 export interface MainLayoutProps {
-  mobileBottomContent?: NavigationSidebarProps['mobileBottomContent'];
-  hamburgerLabel?: NavigationSidebarProps['hamburgerLabel'];
-  mainMenuItems?: NavigationSidebarProps['mainMenuItems'];
   sidebar?: ReactNode;
   header?: ReactNode;
 }
