--- src/components/v5/frame/NavigationSidebar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebar.tsx';
