--- src/components/v5/frame/NavigationSidebar/partials/NavigationFeedbackWidget/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationFeedbackWidget.tsx';
