--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarButton.tsx';
