--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarLink/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarLink.tsx';
