--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarLinksList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarLinksList.tsx';
