--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarLinksList/types.ts
@@ -1,12 +1,0 @@
-import { type NavigationSidebarLinkProps } from '../NavigationSidebarLink/types.ts';
-
-export interface NavigationSidebarLinksListItem
-  extends Omit<NavigationSidebarLinkProps, 'children'> {
-  key: string;
-  label: string;
-}
-
-export interface NavigationSidebarLinksListProps {
-  items: NavigationSidebarLinksListItem[];
-  className?: string;
-}
