--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarMainMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarMainMenu.tsx';
