--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarMobileContentWrapper/consts.ts
@@ -1,8 +1,0 @@
-export const mobileWrapperAnimation = {
-  hidden: {
-    x: '-100%',
-  },
-  visible: {
-    x: 0,
-  },
-};
