--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarMobileContentWrapper/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarMobileContentWrapper.tsx';
