--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarMobileContentWrapper/types.ts
@@ -1,4 +1,0 @@
-export interface NavigationSidebarMobileContentWrapperProps {
-  mobileBottomContent?: React.ReactNode;
-  isOpen?: boolean;
-}
