--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarSecondLevel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarSecondLevel.tsx';
