--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarThirdLevel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSidebarThirdLevel.tsx';
