--- src/components/v5/frame/NavigationSidebar/partials/NavigationSidebarThirdLevel/types.ts
@@ -1,13 +1,0 @@
-import { type ButtonHTMLAttributes } from 'react';
-
-export interface NavigationSidebarThirdLevelItem
-  extends ButtonHTMLAttributes<HTMLButtonElement> {
-  key: string;
-  label: string;
-  href?: string;
-}
-
-export interface NavigationSidebarThirdLevelProps {
-  items: NavigationSidebarThirdLevelItem[];
-  title?: string;
-}
